added d3 chart files to public/js and setup test charts on handlebar partials
