--- poc/client_side_d3/d3barchart.js
@@ -1,497 +1,0 @@
-data1 = [
-    {
-        "date": "13-Apr-12",
-        "value": 65.23
-    },
-    {
-        "date": "12-Apr-12",
-        "value": 62.77
-    },
-    {
-        "date": "11-Apr-12",
-        "value": 66.2
-    },
-    {
-        "date": "10-Apr-12",
-        "value": 68.44
-    },
-    {
-        "date": "9-Apr-12",
-        "value": 66.23
-    },
-    {
-        "date": "5-Apr-12",
-        "value": 63.68
-    },
-    {
-        "date": "4-Apr-12",
-        "value": 64.31
-    },
-    {
-        "date": "3-Apr-12",
-        "value": 69.32
-    },
-    {
-        "date": "2-Apr-12",
-        "value": 61.63
-    }
-];
-data2 = [
-    {
-        "date": "26-Apr-12",
-        "value": 89.7
-    },
-    {
-        "date": "25-Apr-12",
-        "value": 99
-    },
-    {
-        "date": "24-Apr-12",
-        "value": 90.28
-    },
-    {
-        "date": "23-Apr-12",
-        "value": 106.7
-    },
-    {
-        "date": "20-Apr-12",
-        "value": 94.98
-    },
-    {
-        "date": "19-Apr-12",
-        "value": 85.44
-    },
-    {
-        "date": "18-Apr-12",
-        "value": 73.34
-    },
-    {
-        "date": "17-Apr-12",
-        "value": 53.7
-    },
-    {
-        "date": "16-Apr-12",
-        "value": 50.13
-    },
-    {
-        "date": "13-Apr-12",
-        "value": 65.23
-    },
-    {
-        "date": "12-Apr-12",
-        "value": 62.77
-    },
-    {
-        "date": "11-Apr-12",
-        "value": 66.2
-    },
-    {
-        "date": "10-Apr-12",
-        "value": 68.44
-    },
-    {
-        "date": "9-Apr-12",
-        "value": 66.23
-    },
-    {
-        "date": "5-Apr-12",
-        "value": 63.68
-    },
-    {
-        "date": "4-Apr-12",
-        "value": 64.31
-    },
-    {
-        "date": "3-Apr-12",
-        "value": 69.32
-    },
-    {
-        "date": "2-Apr-12",
-        "value": 61.63
-    }
-];
-data3 = [
-    {
-        "date": "11-May-12",
-        "value": 50.13
-    },
-    {
-        "date": "10-May-12",
-        "value": 99.55
-    },
-    {
-        "date": "8-May-12",
-        "value": 76.86
-    },
-    {
-        "date": "6-May-12",
-        "value": 67.62
-    },
-    {
-        "date": "4-May-12",
-        "value": 64.48
-    },
-    {
-        "date": "2-May-12",
-        "value": 60.98
-    },
-    {
-        "date": "1-May-12",
-        "value": 58.13
-    },
-    {
-        "date": "30-Apr-12",
-        "value": 53.98
-    },
-    {
-        "date": "27-Apr-12",
-        "value": 67
-    },
-    {
-        "date": "26-Apr-12",
-        "value": 89.7
-    },
-    {
-        "date": "25-Apr-12",
-        "value": 99
-    },
-    {
-        "date": "24-Apr-12",
-        "value": 90.28
-    },
-    {
-        "date": "23-Apr-12",
-        "value": 106.7
-    },
-    {
-        "date": "20-Apr-12",
-        "value": 94.98
-    },
-    {
-        "date": "19-Apr-12",
-        "value": 85.44
-    },
-    {
-        "date": "18-Apr-12",
-        "value": 73.34
-    },
-    {
-        "date": "17-Apr-12",
-        "value": 53.7
-    },
-    {
-        "date": "16-Apr-12",
-        "value": 50.13
-    },
-    {
-        "date": "13-Apr-12",
-        "value": 65.23
-    },
-    {
-        "date": "12-Apr-12",
-        "value": 62.77
-    },
-    {
-        "date": "11-Apr-12",
-        "value": 66.2
-    },
-    {
-        "date": "10-Apr-12",
-        "value": 68.44
-    },
-    {
-        "date": "9-Apr-12",
-        "value": 66.23
-    },
-    {
-        "date": "5-Apr-12",
-        "value": 63.68
-    },
-    {
-        "date": "4-Apr-12",
-        "value": 64.31
-    },
-    {
-        "date": "3-Apr-12",
-        "value": 69.32
-    },
-    {
-        "date": "2-Apr-12",
-        "value": 61.63
-    }
-];
-data4 = [
-    {
-        "date": "27-May-12",
-        "value": 150.55
-    },    
-    {
-        "date": "27-May-12",
-        "value": 140.55
-    },    
-    {
-        "date": "25-May-12",
-        "value": 145.55
-    },    
-    {
-        "date": "22-May-12",
-        "value": 150.55
-    },
-    {
-        "date": "20-May-12",
-        "value": 160.55
-    },
-    {
-        "date": "18-May-12",
-        "value": 120.86
-    },
-    {
-        "date": "16-May-12",
-        "value": 85.62
-    },
-    {
-        "date": "14-May-12",
-        "value": 65.48
-    },
-    {
-        "date": "12-May-12",
-        "value": 55.98
-    },
-    {
-        "date": "11-May-12",
-        "value": 50.13
-    },
-    {
-        "date": "10-May-12",
-        "value": 99.55
-    },
-    {
-        "date": "8-May-12",
-        "value": 76.86
-    },
-    {
-        "date": "6-May-12",
-        "value": 67.62
-    },
-    {
-        "date": "4-May-12",
-        "value": 64.48
-    },
-    {
-        "date": "2-May-12",
-        "value": 60.98
-    },
-    {
-        "date": "1-May-12",
-        "value": 58.13
-    },
-    {
-        "date": "30-Apr-12",
-        "value": 53.98
-    },
-    {
-        "date": "27-Apr-12",
-        "value": 67
-    },
-    {
-        "date": "26-Apr-12",
-        "value": 89.7
-    },
-    {
-        "date": "25-Apr-12",
-        "value": 99
-    },
-    {
-        "date": "24-Apr-12",
-        "value": 90.28
-    },
-    {
-        "date": "23-Apr-12",
-        "value": 106.7
-    },
-    {
-        "date": "20-Apr-12",
-        "value": 94.98
-    },
-    {
-        "date": "19-Apr-12",
-        "value": 85.44
-    },
-    {
-        "date": "18-Apr-12",
-        "value": 73.34
-    },
-    {
-        "date": "17-Apr-12",
-        "value": 53.7
-    },
-    {
-        "date": "16-Apr-12",
-        "value": 50.13
-    },
-    {
-        "date": "13-Apr-12",
-        "value": 65.23
-    },
-    {
-        "date": "12-Apr-12",
-        "value": 62.77
-    },
-    {
-        "date": "11-Apr-12",
-        "value": 66.2
-    },
-    {
-        "date": "10-Apr-12",
-        "value": 68.44
-    },
-    {
-        "date": "9-Apr-12",
-        "value": 66.23
-    },
-    {
-        "date": "5-Apr-12",
-        "value": 63.68
-    },
-    {
-        "date": "4-Apr-12",
-        "value": 64.31
-    },
-    {
-        "date": "3-Apr-12",
-        "value": 69.32
-    },
-    {
-        "date": "2-Apr-12",
-        "value": 61.63
-    }
-];
-
-//var d3 = require("./greetings.js");
-
-
-var parseDate = d3.time.format("%d-%b-%y").parse;
-
-// set graph dimensions
-var margin = {top: 30, right: 20, bottom: 80, left: 50},
-    width = 1200 - margin.left - margin.right,
-    height = 400 - margin.top - margin.bottom;
-
-var x;
-var y;
-var xAxis;
-var yAxis;
-var svg;
-
-function createBarChart(data){
-
-    x = d3.scale.ordinal()
-        .rangeRoundBands([0, width], .1);
-        
-    y = d3.scale.linear()
-        .range([height, 0]);
-
-    // D3 Axis - renders a d3 scale in SVG
-    xAxis = d3.svg.axis()
-        .scale(x)
-        .orient("bottom")
-        .tickFormat(d3.time.format("%Y-%m-%d"));
-
-    yAxis = d3.svg.axis()
-        .scale(y)
-        .orient("left")
-        .ticks(10, "%");
-
-    // create an SVG element (appended to body)
-    // set size
-    // add a "g" element (think "group")
-    // annoying d3 gotcha - the 'svg' variable here is a 'g' element
-    // the final line sets the transform on <g>, not on <svg>
-    svg = d3.select("body").append("svg")
-        .attr("width", width + margin.left + margin.right)
-        .attr("height", height + margin.top + margin.bottom)
-    .append("g")
-        .attr("transform", "translate(" + margin.left + "," + margin.top + ")");
-
-    svg.append("g")
-        .attr("class", "x axis")
-        .attr("transform", "translate(0," + height + ")")
-
-    svg.append("g")
-        .attr("class", "y axis")
-    .append("text") // just for the title (ticks are automatic)
-        .attr("transform", "rotate(-90)") // rotate the text!
-        .attr("y", 6)
-        .attr("dy", ".71em")
-        .style("text-anchor", "end")
-        .text("Value ($)");
-    
-    updateBarChart(data);
-}
-
-function updateBarChart(data) {
-    data.forEach(function(d) {
-        d.date = parseDate(d.date);
-        d.value = +d.value;
-    });
-    // measure the domain (for x, unique letters) (for y [0,maxFrequency])
-    // now the scales are finished and usable
-    x.domain(data.map(function(d) { return d.date; }));
-    y.domain([0, d3.max(data, function(d) { return d.value; })]);
-
-    // another g element, this time to move the origin to the bottom of the svg element
-    // someSelection.call(thing) is roughly equivalent to thing(someSelection[i])
-    //   for everything in the selection\
-    // the end result is g populated with text and lines!
-    if(data.length<12){
-        svg.select('.x.axis').transition().duration(300).call(xAxis)
-    }else{
-        svg.select('.x.axis').transition().duration(300).call(xAxis)
-            .selectAll("text")
-                .attr("y", 0)
-                .attr("x", 9)
-                .attr("dy", ".35em")
-                .attr("transform", "rotate(90)")
-                .style("text-anchor", "start");
-    }
-
-
-    // same for yAxis but with more transform and a title
-    svg.select(".y.axis").transition().duration(300).call(yAxis)
-
-    // THIS IS THE ACTUAL WORK!
-    var bars = svg.selectAll(".bar").data(data, function(d) { return d.date; }) // (data) is an array/iterable thing, second argument is an ID generator function
-
-    bars.exit()
-        .transition()
-        .duration(300)
-        .attr("y", y(0))
-        .attr("height", height - y(0))
-        .style('fill-opacity', 1e-6)
-        .remove();
-
-    // data that needs DOM = enter() (a set/selection, not an event!)
-    bars.enter().append("rect")
-        .attr("class", "bar")
-        .style("fill", "steelblue")
-        .attr("y", y(0))
-        .attr("height", height - y(0));
-
-    // the "UPDATE" set:
-    bars.transition().duration(300).attr("x", function(d) { return x(d.date); }) // (d) is one item from the data array, x is the scale object from above
-        .attr("width", x.rangeBand()) // constant, so no callback function(d) here
-        .style("fill", "steelblue")
-        .attr("y", function(d) { return y(d.value); })
-        .attr("height", function(d) { return height - y(d.value); }); // flip the height, because y's domain is bot
-  }
-
-createBarChart(data1);
-setTimeout(function(){
-    updateBarChart(data2);
-}, 5000);
-setTimeout(function(){
-    updateBarChart(data3);
-}, 10000);
-setTimeout(function(){
-    updateBarChart(data4);
-}, 15000);
