--- poc/client_side_d3/d3linechart.js
@@ -1,503 +1,0 @@
-data1 = [
-    {
-        "xprop": "13-Apr-12",
-        "yprop": 65.23
-    },
-    {
-        "xprop": "12-Apr-12",
-        "yprop": 62.77
-    },
-    {
-        "xprop": "11-Apr-12",
-        "yprop": 66.2
-    },
-    {
-        "xprop": "10-Apr-12",
-        "yprop": 68.44
-    },
-    {
-        "xprop": "9-Apr-12",
-        "yprop": 66.23
-    },
-    {
-        "xprop": "5-Apr-12",
-        "yprop": 63.68
-    },
-    {
-        "xprop": "4-Apr-12",
-        "yprop": 64.31
-    },
-    {
-        "xprop": "3-Apr-12",
-        "yprop": 69.32
-    },
-    {
-        "xprop": "2-Apr-12",
-        "yprop": 61.63
-    }
-];
-data2 = [
-    {
-        "xprop": "26-Apr-12",
-        "yprop": 89.7
-    },
-    {
-        "xprop": "25-Apr-12",
-        "yprop": 99
-    },
-    {
-        "xprop": "24-Apr-12",
-        "yprop": 90.28
-    },
-    {
-        "xprop": "23-Apr-12",
-        "yprop": 106.7
-    },
-    {
-        "xprop": "20-Apr-12",
-        "yprop": 94.98
-    },
-    {
-        "xprop": "19-Apr-12",
-        "yprop": 85.44
-    },
-    {
-        "xprop": "18-Apr-12",
-        "yprop": 73.34
-    },
-    {
-        "xprop": "17-Apr-12",
-        "yprop": 53.7
-    },
-    {
-        "xprop": "16-Apr-12",
-        "yprop": 50.13
-    },
-    {
-        "xprop": "13-Apr-12",
-        "yprop": 65.23
-    },
-    {
-        "xprop": "12-Apr-12",
-        "yprop": 62.77
-    },
-    {
-        "xprop": "11-Apr-12",
-        "yprop": 66.2
-    },
-    {
-        "xprop": "10-Apr-12",
-        "yprop": 68.44
-    },
-    {
-        "xprop": "9-Apr-12",
-        "yprop": 66.23
-    },
-    {
-        "xprop": "5-Apr-12",
-        "yprop": 63.68
-    },
-    {
-        "xprop": "4-Apr-12",
-        "yprop": 64.31
-    },
-    {
-        "xprop": "3-Apr-12",
-        "yprop": 69.32
-    },
-    {
-        "xprop": "2-Apr-12",
-        "yprop": 61.63
-    }
-];
-data3 = [
-    {
-        "xprop": "11-May-12",
-        "yprop": 50.13
-    },
-    {
-        "xprop": "10-May-12",
-        "yprop": 99.55
-    },
-    {
-        "xprop": "8-May-12",
-        "yprop": 76.86
-    },
-    {
-        "xprop": "6-May-12",
-        "yprop": 67.62
-    },
-    {
-        "xprop": "4-May-12",
-        "yprop": 64.48
-    },
-    {
-        "xprop": "2-May-12",
-        "yprop": 60.98
-    },
-    {
-        "xprop": "1-May-12",
-        "yprop": 58.13
-    },
-    {
-        "xprop": "30-Apr-12",
-        "yprop": 53.98
-    },
-    {
-        "xprop": "27-Apr-12",
-        "yprop": 67
-    },
-    {
-        "xprop": "26-Apr-12",
-        "yprop": 89.7
-    },
-    {
-        "xprop": "25-Apr-12",
-        "yprop": 99
-    },
-    {
-        "xprop": "24-Apr-12",
-        "yprop": 90.28
-    },
-    {
-        "xprop": "23-Apr-12",
-        "yprop": 106.7
-    },
-    {
-        "xprop": "20-Apr-12",
-        "yprop": 94.98
-    },
-    {
-        "xprop": "19-Apr-12",
-        "yprop": 85.44
-    },
-    {
-        "xprop": "18-Apr-12",
-        "yprop": 73.34
-    },
-    {
-        "xprop": "17-Apr-12",
-        "yprop": 53.7
-    },
-    {
-        "xprop": "16-Apr-12",
-        "yprop": 50.13
-    },
-    {
-        "xprop": "13-Apr-12",
-        "yprop": 65.23
-    },
-    {
-        "xprop": "12-Apr-12",
-        "yprop": 62.77
-    },
-    {
-        "xprop": "11-Apr-12",
-        "yprop": 66.2
-    },
-    {
-        "xprop": "10-Apr-12",
-        "yprop": 68.44
-    },
-    {
-        "xprop": "9-Apr-12",
-        "yprop": 66.23
-    },
-    {
-        "xprop": "5-Apr-12",
-        "yprop": 63.68
-    },
-    {
-        "xprop": "4-Apr-12",
-        "yprop": 64.31
-    },
-    {
-        "xprop": "3-Apr-12",
-        "yprop": 69.32
-    },
-    {
-        "xprop": "2-Apr-12",
-        "yprop": 61.63
-    }
-];
-data4 = [
-    {
-        "xprop": "27-May-12",
-        "yprop": 150.55
-    },    
-    {
-        "xprop": "27-May-12",
-        "yprop": 140.55
-    },    
-    {
-        "xprop": "25-May-12",
-        "yprop": 145.55
-    },    
-    {
-        "xprop": "22-May-12",
-        "yprop": 150.55
-    },
-    {
-        "xprop": "20-May-12",
-        "yprop": 160.55
-    },
-    {
-        "xprop": "18-May-12",
-        "yprop": 120.86
-    },
-    {
-        "xprop": "16-May-12",
-        "yprop": 85.62
-    },
-    {
-        "xprop": "14-May-12",
-        "yprop": 65.48
-    },
-    {
-        "xprop": "12-May-12",
-        "yprop": 55.98
-    },
-    {
-        "xprop": "11-May-12",
-        "yprop": 50.13
-    },
-    {
-        "xprop": "10-May-12",
-        "yprop": 99.55
-    },
-    {
-        "xprop": "8-May-12",
-        "yprop": 76.86
-    },
-    {
-        "xprop": "6-May-12",
-        "yprop": 67.62
-    },
-    {
-        "xprop": "4-May-12",
-        "yprop": 64.48
-    },
-    {
-        "xprop": "2-May-12",
-        "yprop": 60.98
-    },
-    {
-        "xprop": "1-May-12",
-        "yprop": 58.13
-    },
-    {
-        "xprop": "30-Apr-12",
-        "yprop": 53.98
-    },
-    {
-        "xprop": "27-Apr-12",
-        "yprop": 67
-    },
-    {
-        "xprop": "26-Apr-12",
-        "yprop": 89.7
-    },
-    {
-        "xprop": "25-Apr-12",
-        "yprop": 99
-    },
-    {
-        "xprop": "24-Apr-12",
-        "yprop": 90.28
-    },
-    {
-        "xprop": "23-Apr-12",
-        "yprop": 106.7
-    },
-    {
-        "xprop": "20-Apr-12",
-        "yprop": 94.98
-    },
-    {
-        "xprop": "19-Apr-12",
-        "yprop": 85.44
-    },
-    {
-        "xprop": "18-Apr-12",
-        "yprop": 73.34
-    },
-    {
-        "xprop": "17-Apr-12",
-        "yprop": 53.7
-    },
-    {
-        "xprop": "16-Apr-12",
-        "yprop": 50.13
-    },
-    {
-        "xprop": "13-Apr-12",
-        "yprop": 65.23
-    },
-    {
-        "xprop": "12-Apr-12",
-        "yprop": 62.77
-    },
-    {
-        "xprop": "11-Apr-12",
-        "yprop": 66.2
-    },
-    {
-        "xprop": "10-Apr-12",
-        "yprop": 68.44
-    },
-    {
-        "xprop": "9-Apr-12",
-        "yprop": 66.23
-    },
-    {
-        "xprop": "5-Apr-12",
-        "yprop": 63.68
-    },
-    {
-        "xprop": "4-Apr-12",
-        "yprop": 64.31
-    },
-    {
-        "xprop": "3-Apr-12",
-        "yprop": 69.32
-    },
-    {
-        "xprop": "2-Apr-12",
-        "yprop": 61.63
-    }
-];
-
-//var d3 = require("./greetings.js");
-var config = {
-    margin: {top: 30, right: 20, bottom: 30, left: 50},
-    width: 375,
-    height: 375,
-    timeSeries: false,
-    el: "body"
-};
-
-
-// data/time parsing
-var parseDate = d3.time.format("%d-%b-%y").parse;
-
-// Define the line
-var valueline;
-
-// set graph dimensions
-var margin = {top: 30, right: 20, bottom: 30, left: 50},
-    width = 1200 - margin.left - margin.right,
-    height = 400 - margin.top - margin.bottom;
-
-var x;
-var y;
-var xAxis;
-var yAxis;
-
-var lines = [];
-
-function createLineChart(data){
-
-    // set the ranges by graph dimension
-    if(config.timeSeries){
-        x = d3.time.scale().range([0, width]);
-    }else{
-        x = d3.scale.linear().range([0, width]);
-    }
-    
-    
-    y = d3.scale.linear().range([height, 0]);
-
-    // define graph x/y axis
-    xAxis = d3.svg.axis().scale(x)
-        .orient("bottom")
-    if(config.timeSeries){
-        xAxis.ticks(5)
-            .tickFormat(d3.time.format("%Y-%m-%d"))
-    }else{
-        xAxis.ticks(data.length);
-    }
-
-    yAxis = d3.svg.axis().scale(y)
-        .orient("left").ticks(5);
-
-    valueline = d3.svg.line()
-        .x(function(d) { return x(d.xprop); })
-        .y(function(d) { return y(d.yprop); });        
-        
-    // Adds the svg canvas
-    var svg = d3.select("body")
-        .append("svg")
-            .attr("width", width + margin.left + margin.right)
-            .attr("height", height + margin.top + margin.bottom)
-        .append("g")
-            .attr("transform", 
-                "translate(" + margin.left + "," + margin.top + ")");
-
-    data.forEach(function(d) {
-        if(config.timeSeries){
-            d.xprop = parseDate(d.xprop); 
-        }
-        d.yprop = +d.yprop;
-    });
-
-    // Scale the range of the data
-    x.domain(d3.extent(data, function(d) { return d.xprop; }));
-    y.domain([0, d3.max(data, function(d) { return d.yprop; })]);
-
-    // Add the valueline path.
-    svg.append("path")
-        .attr("class", "line")
-        .attr("d", valueline(data));
-
-    // Add the X Axis
-    svg.append("g")
-        .attr("class", "x axis")
-        .attr("transform", "translate(0," + height + ")")
-        .call(xAxis);
-
-    // Add the Y Axis
-    svg.append("g")
-        .attr("class", "y axis")
-        .call(yAxis);
-};
-
-datanottime = [
-    {
-        "xprop": 1,
-        "yprop": 1
-    },
-    {
-        "xprop": 2,
-        "yprop": 1
-    },
-    {
-        "xprop": 3,
-        "yprop": 2
-    },
-    {
-        "xprop": 4,
-        "yprop": 3
-    },
-    {
-        "xprop": 5,
-        "yprop": 5
-    },
-    {
-        "xprop": 6,
-        "yprop": 8
-    },
-    {
-        "xprop": 7,
-        "yprop": 13
-    },
-    {
-        "xprop": 8,
-        "yprop": 21
-    },
-    {
-        "xprop": 9,
-        "yprop": 34
-    }
-];
-
-createLineChart(datanottime);
