+++ views/index.handlebars
@@ -2,3 +2,8 @@
 
 <h2>Hello World</h2>
 <p>Lorem ipsum dolor, sit amet consectetur adipisicing elit. Eum nihil deleniti magni necessitatibus aliquid eius, eos pariatur minus nemo ipsa ab, soluta cumque voluptatibus rem, debitis sequi hic quis? Laboriosam.</p>
+
+{{> linechart}}
+{{> barchart}}
+{{> donutchart}}
+{{> piechart}}
